In this commit the following high level functions have been modified and added:
    - Declare page tables for *most* user pages from FreeBSD
    - Use sva_update_l*_mapping for managing the initial access for the newly
      created page as opposed to allowing FreeBSD to directly create the
      mapping.

Note that in this update, to the FreeBSD code I have added, in pmaps.c, to
declarations: SVA_ENA and SVA_ENA_DEBUG, to control whether or not SVA
intrinsics will be called to manage paging. By turning off the SVA_ENA function
the code will use all the previous methods for managing paging. This allows us
to keep the old code and easily add in switches for going between the original
and SVA versions of the code. 

The updates being established via the intrinsics (sva_update_l*_mapping) is
only setting up the default mapping without any validation. 

In this update I also add the declare and update to the l4 pages, as well as
capture the intialization of several (maybe all) l4 user space pages upon
process creation. This must be investigated further, but it does show that the
declare and updates are being used correctly for l1 - l4. 



git-svn-id: https://subversion.cs.illinois.edu/svn/llvm_group/sva/trunk@11327 42e9315a-1956-0410-9e23-90ee2fe044bf