This commit represents an initial effort to add in verification routines to page table updates. These updates caused a set of other changes, such as modifications to update functions and declare routines.
Another major change in this commit is that I have moved the sva_* mmu
intrinsics functions to a new header file just for the mmu intrinsics
functions. I did this in an effort to eliminate the need to include mmu.h in
the kernel in order to avoid name clashes, to reduce the compilation time on a
modification to mmu.h, and to more cleanly separate the two different pieces of
functionality. 

In init.c I add in the call to the mmu_init function, however, this design will
be changing as the new initialization method is added in a future commit. 

In making changes for the verification functionality I moved several things
into mmu.h to make things a bit cleaner. 



git-svn-id: https://subversion.cs.illinois.edu/svn/llvm_group/sva/trunk@11393 42e9315a-1956-0410-9e23-90ee2fe044bf