The major change in this update is that we are now setting the write protect flag in SVA mmu updates, and we have also enabled the read only flag for pml4 pages.
In testing the write protect a bug popped up when the kernel ran the function
pmap_qremove. The solution was to modify the underlying page table update
function, which I replaced with an SVA function. 

The new function is page table page level agnostic SVA intrinsic to update an
mmu mapping. This new function can be used for all updates, but for not is just
enabled for this particular location. 

There are also a few refactorings done in mmu.c and mmu.h to dynamically handle
setting write protect for a page table entry. 



git-svn-id: https://subversion.cs.illinois.edu/svn/llvm_group/sva/trunk@11500 42e9315a-1956-0410-9e23-90ee2fe044bf