Modified the page reference count checks to allow for multiple references to page table pages, but only on VA address pointed at the pte.
This necessitated some changes in the way we manage metadata on updates, which
is currently commented out but has the framework in the code.

There is a bug that this commit also doesn't exercise, which is to update
metadata for the physical address 0 (page 0). I am currently working on finding
out what the issue is.



git-svn-id: https://subversion.cs.illinois.edu/svn/llvm_group/sva/trunk@11551 42e9315a-1956-0410-9e23-90ee2fe044bf