Established a default secret key creation for every process/thread that is created on the system.
Upon initializing the sva thread data structure a key is added to a new secret
variable conatining the thread specific secret key. 

This is just an initial version as we will need to dynamically obtain the key
from the executable for a real solution, but this should enable us to use the
key now for usage by the application.



git-svn-id: https://subversion.cs.illinois.edu/svn/llvm_group/sva/trunk@11773 42e9315a-1956-0410-9e23-90ee2fe044bf