Restore UMA_MD_SMALL_ALLOC for potential performance impact avoidance.
Disabling it (as done in a previous commit) stops the generation
of mapping requests that SVA has to modify from RW to RO,
but there's a reason this is done this way and for the eval
avoiding changing this sort of thing is best.

Restoring it causes SVA to warn about having to change
the requests mappings to be read-only, but the recent
shrinking of sva's preallocated datastructures seems
to be enough to get things booting in a VM for me.