Ensure WP bit is set on traps/interrupts/syscalls.
This isn't done for common case--since we're
coming from untrusted code it will already be set.

However, in theory one could jump to a "clear WP"
write and cause an interrupt to immediately occur,
transferring control to their handler with
access to protected memory.

To foil such an attack, we guarantee we control
traps/interrupts/syscalls go to us first
where we set the WP bit as in this commit.