Added new cr3 update approach: just to simulate perf.
The new approach is to map out the page with the real write to cr3 instruciton
that way an attacker cannot ever direclty target it. There was a problem under
the old approach in that the attacker could directly modify the cr3, which
would change the interpretation of the address comparison to verify the new
value is a declared L4 page. The new approach: 1) unsets cr0, 2) adds the
mapping (or transitions it from NX to EX), 3) flushes tlb, 4) jumps to the
funciton, 5) does check and update, 6) jumps back and unmaps the mapping (or
marks it nx), 7) reenable wp=1.